Set the correct page for the selected live image (unfortunately, EventDisplayViewController loads images in arbitrary order, many thanks to E.F.)
git-svn-id: http://root.cern.ch/svn/cern-app/trunk@99 59dca7bf-9055-4625-9011-bab61c547c97

+++ Classes/Utilities/ContentProviders.h
@@ -48,6 +48,6 @@
 
 - (void) addPageWithContentTo : (MultiPageController *) controller;
 - (void) loadControllerTo : (UINavigationController *) controller;
-- (void) loadControllerTo : (UINavigationController *) controller selectedImage : (unsigned) selected;
+- (void) loadControllerTo : (UINavigationController *) controller selectedImage : (NSInteger) selected;
 
 @end
